Address PR review: cache VoidResponse and fix LongMessage roundtrip
- VoidResponseSerializer.DeserializeCore now returns VoidResponse.CachedInstance to avoid allocating a new instance per response.

- LongMessageSerializer.DeserializeCore now reads via ReadLong to match the WriteLong wire format used by SerializeCore.

+++ test/UnitTests/Microsoft.Testing.Platform.UnitTests/IPC/IPCTests.cs
@@ -268,7 +268,7 @@ public sealed class IPCTests
     {
         public override int Id => 4;
 
-        protected override LongMessage DeserializeCore(Stream stream) => new(ReadInt(stream));
+        protected override LongMessage DeserializeCore(Stream stream) => new(ReadLong(stream));
 
         protected override void SerializeCore(LongMessage objectToSerialize, Stream stream) => WriteLong(stream, objectToSerialize.Long);
     }
